+++ src/main/java/com/example/bo/CardHolder.java
@@ -32,4 +32,11 @@ public class CardHolder {
     private String modifiedBy;
 
     private CorporateMetaData metaData;
+
+    public void update() {
+        if(this.getModifiedBy()==null || this.getModifiedBy().trim().length()==0) {
+            this.setModifiedBy("SYSTEM");
+        }
+        this.setModifiedDate(new Date());
+    }
 }