+++ src/main/java/com/example/interceptor/LoggerInterceptor.java
@@ -1,0 +1,17 @@
+package com.example.interceptor;
+
+import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class LoggerInterceptor extends HandlerInterceptorAdapter {
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response,
+                             Object handler) {
+
+        System.out.println("[*][*] Request received from :: " + request.getRemoteAddr());
+        return true;
+    }
+}